feat: Move ItemUsageService to ItemRelationsService
