--- model/SearchEngine/Service/ItemUsageService.php
@@ -1,47 +1,0 @@
-<?php
-
-/**
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; under version 2
- * of the License (non-upgradable).
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- *
- * Copyright (c) 2025 (original work) Open Assessment Technologies SA;
- */
-
-declare(strict_types=1);
-
-namespace oat\taoAdvancedSearch\model\SearchEngine\Service;
-
-use oat\tao\model\search\ResultSet;
-use oat\taoAdvancedSearch\model\SearchEngine\Driver\Elasticsearch\ElasticSearch;
-use oat\taoAdvancedSearch\model\SearchEngine\Query;
-
-class ItemUsageService
-{
-    private const TEST_INDEX = 'tests';
-    private ElasticSearch $elasticSearch;
-
-    public function __construct(ElasticSearch $elasticSearch)
-    {
-        $this->elasticSearch = $elasticSearch;
-    }
-
-    public function getItemTests(array $itemUris): ResultSet
-    {
-        $query = new Query(self::TEST_INDEX);
-        foreach ($itemUris as $itemUri) {
-            $query->addCondition(sprintf('item_uris:"%s"', $itemUri));
-        }
-        return $this->elasticSearch->search($query);
-    }
-}
